use HMAC instead of a straight sha1 of the content
I think this is better.  What we now do is AES encrypt a 1-block salt
with the session data, append the IV & compute the HMAC-SHA256 of that
(ciphertext + IV).  Finally we append the HMAC to the cipertext + IV,
and base64 encode the whole thing to make it storeable in a cookie.